+++ src/components/App/Header/MailboxContent.module.css
@@ -1,0 +1,53 @@
+@value main-black, border-gray, yellow-bar from "../../common/variables.module.css";
+
+.mailboxContainer {
+  overflow-y: auto;
+  line-height: 1.5;
+
+  .header {
+    border-bottom: 1px solid border-gray;
+    padding: 12px;
+
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+
+    .title {
+      font-size: 16px;
+      font-weight: 600;
+    }
+
+    .buttons {
+      display: flex;
+      gap: 8px;
+    }
+  }
+
+  .date {
+    font-size: 11px;
+    color: #9a9a9a;
+  }
+
+  .button {
+    border: 1px solid #9a9a9a;
+    border-radius: 45px;
+    font-size: 14px;
+    font-weight: 400;
+    line-height: 1.5;
+    padding: 0 12px;
+    color: #9a9a9a;
+
+    &.active {
+      background-color: yellow-bar;
+      border: 1px solid yellow-bar;
+      color: main-black;
+      font-weight: 500;
+    }
+  }
+
+  .loadMore {
+    display: flex;
+    justify-content: center;
+    padding: 0 12px 12px;
+  }
+}
